fix: compare result of `G_N_ELEMENTS()` with `size_t`
I've been told the actual result of `G_N_ELEMENTS()` is actually
`size_t`, rather than e.g. `unsigned int`, so fix all instances of
the incorrect comparison.

Additionally, don't use a static variable to store the length of
test arrays, since `G_N_ELEMENTS()` is a pre-processor macro anyways.

+++ src/libvalent/device/valent-device-manager.c
@@ -796,7 +796,7 @@ valent_device_manager_shutdown (ValentApplicationPlugin *plugin)
 
       if (application != NULL)
         {
-          for (unsigned int i = 0; i < G_N_ELEMENTS (app_actions); i++)
+          for (size_t i = 0; i < G_N_ELEMENTS (app_actions); i++)
             g_action_map_remove_action (G_ACTION_MAP (application),
                                         app_actions[i].name);
